requirements.txt
# ============================================================
# SMB-in-a-Box / T-CHai  —  Python Dependencies
# Install: pip install -r requirements.txt
# ============================================================

# --- Web framework ---
fastapi==0.111.1
uvicorn[standard]==0.30.1

# --- AI / Agents ---
anthropic==0.28.0

# --- Configuration & validation ---
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1

# --- Database ---
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.2

# --- HTTP client ---
httpx==0.27.0

# --- SMS / Communications ---
twilio==9.2.3

# --- Email ---
sendgrid==6.11.0

# --- Google APIs ---
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0

# --- Utilities ---
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# --- Testing ---
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-httpx==0.30.0
respx==0.21.1

# --- Dev / Linting ---
ruff==0.5.3
mypy==1.11.0
